# specs/openapi/master-data.v1.yaml
openapi: 3.1.0
info:
  title: HIMS Master Data API
  version: 1.0.0
  description: 'HIMS Master Data catalog API — platform reference data (modules first; more surfaces phased in). Base path
    `/api/v1/master-data`. DELETE on modules is a recursive soft-delete (`is_deleted = true`). Gateway-level authentication
    (API keys / JWT) is expected before production; this spec does not model per-route security yet.

    '
servers:
- url: /api/v1/master-data
paths:
...
  /visitpad/chief-complaints:
    get:
      tags:
      - Visitpad — Chief complaints
      summary: List chief complaints
      operationId: get_chief_complaints_api_v1_master_data_visitpad_chief_complaints_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: body_system
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/VisitpadBodySystem'
          - type: 'null'
          title: Body System
      - name: triage_priority
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/VisitpadTriagePriority'
          - type: 'null'
          title: Triage Priority
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadChiefComplaintListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Visitpad — Chief complaints
      summary: Create chief complaint
      operationId: post_chief_complaint_api_v1_master_data_visitpad_chief_complaints_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitpadChiefComplaintCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadChiefComplaintSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /visitpad/chief-complaints/{chief_complaint_id}:
    get:
      tags:
      - Visitpad — Chief complaints
      summary: Get chief complaint
      operationId: get_chief_complaint_api_v1_master_data_visitpad_chief_complaints__chief_complaint_id__get
      parameters:
      - name: chief_complaint_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Chief Complaint Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadChiefComplaintSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Visitpad — Chief complaints
      summary: Update chief complaint
      operationId: patch_chief_complaint_api_v1_master_data_visitpad_chief_complaints__chief_complaint_id__patch
      parameters:
      - name: chief_complaint_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Chief Complaint Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitpadChiefComplaintUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadChiefComplaintSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Visitpad — Chief complaints
      summary: Soft-delete chief complaint
      operationId: delete_chief_complaint_api_v1_master_data_visitpad_chief_complaints__chief_complaint_id__delete
      parameters:
      - name: chief_complaint_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Chief Complaint Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadChiefComplaintSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /visitpad/chronic-illnesses:
    get:
      tags:
      - Visitpad — Chronic illnesses
      summary: List chronic illnesses
      operationId: get_chronic_illnesses_api_v1_master_data_visitpad_chronic_illnesses_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/VisitpadChronicIllnessCategory'
          - type: 'null'
          title: Category
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadChronicIllnessListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Visitpad — Chronic illnesses
      summary: Create chronic illness
      operationId: post_chronic_illness_api_v1_master_data_visitpad_chronic_illnesses_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitpadChronicIllnessCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadChronicIllnessSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /visitpad/chronic-illnesses/{chronic_illness_id}:
    get:
      tags:
      - Visitpad — Chronic illnesses
      summary: Get chronic illness
      operationId: get_chronic_illness_api_v1_master_data_visitpad_chronic_illnesses__chronic_illness_id__get
      parameters:
      - name: chronic_illness_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Chronic Illness Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadChronicIllnessSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Visitpad — Chronic illnesses
      summary: Update chronic illness
      operationId: patch_chronic_illness_api_v1_master_data_visitpad_chronic_illnesses__chronic_illness_id__patch
      parameters:
      - name: chronic_illness_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Chronic Illness Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitpadChronicIllnessUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadChronicIllnessSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Visitpad — Chronic illnesses
      summary: Soft-delete chronic illness
      operationId: delete_chronic_illness_api_v1_master_data_visitpad_chronic_illnesses__chronic_illness_id__delete
      parameters:
      - name: chronic_illness_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Chronic Illness Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadChronicIllnessSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VisitpadBodySystem:
      type: string
      enum:
      - cardiovascular
      - respiratory
      - neurological
      - gastrointestinal
      - musculoskeletal
      - ent
      - skin
      - psychiatric
      - other
      title: VisitpadBodySystem
    VisitpadChiefComplaintCreate:
      properties:
        code:
          type: string
          maxLength: 64
          minLength: 1
          title: Code
        display_name:
          type: string
          maxLength: 256
          minLength: 1
          title: Display Name
        body_system:
          $ref: '#/components/schemas/VisitpadBodySystem'
        triage_priority:
          $ref: '#/components/schemas/VisitpadTriagePriority'
        synonyms:
          items:
            type: string
          type: array
          maxItems: 50
          title: Synonyms
        is_paediatric_relevant:
          type: boolean
          title: Is Paediatric Relevant
          default: false
        display_order:
          type: integer
          title: Display Order
          default: 0
        is_active:
          type: boolean
          title: Is Active
          default: true
        snomed_code:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Snomed Code
      additionalProperties: false
      type: object
      required:
      - code
      - display_name
      - body_system
      - triage_priority
      title: VisitpadChiefComplaintCreate
    VisitpadChiefComplaintListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/VisitpadChiefComplaintResponse'
          type: array
          title: Data
        total:
          type: integer
          title: Total
      type: object
      required:
      - data
      - total
      title: VisitpadChiefComplaintListResponse
    VisitpadChiefComplaintResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        tenant_id:
          type: string
          format: uuid
          title: Tenant Id
        code:
          type: string
          title: Code
        display_name:
          type: string
          title: Display Name
        body_system:
          $ref: '#/components/schemas/VisitpadBodySystem'
        triage_priority:
          $ref: '#/components/schemas/VisitpadTriagePriority'
        synonyms:
          items:
            type: string
          type: array
          title: Synonyms
        is_paediatric_relevant:
          type: boolean
          title: Is Paediatric Relevant
        display_order:
          type: integer
          title: Display Order
        is_active:
          type: boolean
          title: Is Active
        is_deleted:
          type: boolean
          title: Is Deleted
        snomed_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Snomed Code
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - tenant_id
      - code
      - display_name
      - body_system
      - triage_priority
      - synonyms
      - is_paediatric_relevant
      - display_order
      - is_active
      - is_deleted
      - created_at
      - updated_at
      title: VisitpadChiefComplaintResponse
    VisitpadChiefComplaintSingleResponse:
      properties:
        data:
          $ref: '#/components/schemas/VisitpadChiefComplaintResponse'
      type: object
      required:
      - data
      title: VisitpadChiefComplaintSingleResponse
    VisitpadChiefComplaintUpdate:
      properties:
        code:
          anyOf:
          - type: string
            maxLength: 64
            minLength: 1
          - type: 'null'
          title: Code
        display_name:
          anyOf:
          - type: string
            maxLength: 256
            minLength: 1
          - type: 'null'
          title: Display Name
        body_system:
          anyOf:
          - $ref: '#/components/schemas/VisitpadBodySystem'
          - type: 'null'
        triage_priority:
          anyOf:
          - $ref: '#/components/schemas/VisitpadTriagePriority'
          - type: 'null'
        synonyms:
          anyOf:
          - items:
              type: string
            type: array
            maxItems: 50
          - type: 'null'
          title: Synonyms
        is_paediatric_relevant:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Paediatric Relevant
        display_order:
          anyOf:
          - type: integer
          - type: 'null'
          title: Display Order
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        snomed_code:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Snomed Code
        is_deleted:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Deleted
      additionalProperties: false
      type: object
      title: VisitpadChiefComplaintUpdate
    VisitpadChronicIllnessCategory:
      type: string
      enum:
      - cardiovascular
      - respiratory
      - metabolic
      - renal
      - neurological
      - other
      title: VisitpadChronicIllnessCategory
    VisitpadChronicIllnessCreate:
      properties:
        display_name:
          type: string
          maxLength: 512
          minLength: 1
          title: Display Name
        icd10_code:
          type: string
          maxLength: 16
          minLength: 1
          title: Icd10 Code
        category:
          $ref: '#/components/schemas/VisitpadChronicIllnessCategory'
        snomed_code:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Snomed Code
        display_order:
          type: integer
          title: Display Order
          default: 0
        is_active:
          type: boolean
          title: Is Active
          default: true
      additionalProperties: false
      type: object
      required:
      - display_name
      - icd10_code
      - category
      title: VisitpadChronicIllnessCreate
    VisitpadChronicIllnessListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/VisitpadChronicIllnessResponse'
          type: array
          title: Data
        total:
          type: integer
          title: Total
      type: object
      required:
      - data
      - total
      title: VisitpadChronicIllnessListResponse
    VisitpadChronicIllnessResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        tenant_id:
          type: string
          format: uuid
          title: Tenant Id
        display_name:
          type: string
          title: Display Name
        icd10_code:
          type: string
          title: Icd10 Code
        category:
          $ref: '#/components/schemas/VisitpadChronicIllnessCategory'
        snomed_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Snomed Code
        display_order:
          type: integer
          title: Display Order
        is_active:
          type: boolean
          title: Is Active
        is_deleted:
          type: boolean
          title: Is Deleted
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - tenant_id
      - display_name
      - icd10_code
      - category
      - display_order
      - is_active
      - is_deleted
      - created_at
      - updated_at
      title: VisitpadChronicIllnessResponse
    VisitpadChronicIllnessSingleResponse:
      properties:
        data:
          $ref: '#/components/schemas/VisitpadChronicIllnessResponse'
      type: object
      required:
      - data
      title: VisitpadChronicIllnessSingleResponse
    VisitpadChronicIllnessUpdate:
      properties:
        display_name:
          anyOf:
          - type: string
            maxLength: 512
            minLength: 1
          - type: 'null'
          title: Display Name
        icd10_code:
          anyOf:
          - type: string
            maxLength: 16
            minLength: 1
          - type: 'null'
          title: Icd10 Code
        category:
          anyOf:
          - $ref: '#/components/schemas/VisitpadChronicIllnessCategory'
          - type: 'null'
        snomed_code:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Snomed Code
        display_order:
          anyOf:
          - type: integer
          - type: 'null'
          title: Display Order
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        is_deleted:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Deleted
      additionalProperties: false
      type: object
      title: VisitpadChronicIllnessUpdate
...
    VisitpadTriagePriority:
      type: string
      enum:
      - urgent
      - semi_urgent
      - non_urgent
      - routine
      title: VisitpadTriagePriority
...